+++ admin/tabs/general/alertBanner/AlertBannerPanel.ts
@@ -37,7 +37,7 @@ import {toolbar} from '@xh/hoist/desktop/cmp/toolbar';
 import {dateTimeRenderer} from '@xh/hoist/format';
 import {Icon} from '@xh/hoist/icon';
 import {menu, menuItem, popover} from '@xh/hoist/kit/blueprint';
-import {LocalDate, SECONDS} from '@xh/hoist/utils/datetime';
+import {LocalDate} from '@xh/hoist/utils/datetime';
 import {isEmpty} from 'lodash';
 import {ReactNode} from 'react';
 import {AlertBannerModel} from './AlertBannerModel';
@@ -72,15 +72,13 @@ const formPanel = hoistCmp.factory<AlertBannerModel>(({model}) => {
                 labelWidth: 100
             },
             items: [
-                XH.alertBannerService.enabled
+                XH.getConf('xhAlertBannerConfig', {}).enabled
                     ? div({
                           className: `${baseClassName}__intro`,
                           items: [
                               p(`Show an alert banner to all ${XH.appName} users.`),
                               p(
-                                  `Configure and preview below. Presets can be saved and loaded via bottom bar menu. Banner will appear to all users within ${
-                                      XH.alertBannerService.interval / SECONDS
-                                  }s once marked Active and saved.`
+                                  'Configure and preview below. Presets can be saved and loaded via bottom bar menu. Banner will appear to all users once marked Active and saved.'
                               )
                           ]
                       })