+++ svc/AlertBannerService.ts
@@ -6,52 +6,28 @@
  */
 import {BannerModel} from '@xh/hoist/appcontainer/BannerModel';
 import {markdown} from '@xh/hoist/cmp/markdown';
-import {BannerSpec, HoistService, Intent, managed, XH} from '@xh/hoist/core';
+import {BannerSpec, HoistService, Intent, XH} from '@xh/hoist/core';
 import {Icon} from '@xh/hoist/icon';
-import {Timer} from '@xh/hoist/utils/async';
-import {SECONDS} from '@xh/hoist/utils/datetime';
 import {compact, isEmpty, map, trim} from 'lodash';
 
 /**
  * Service to display an app-wide alert banner, as configured via the Hoist Admin console.
  *
- * For this service to be active, a client-visible `xhAlertBannerConfig` config must be specified
- * as `{enabled:true, interval: x}`, where `x` sets this service's polling frequency in seconds.
+ * Note that the client is provided with updated banner data from the server via
+ * EnvironmentService, and its regular polling.  See 'xhEnvPollConfig' for more information.
  */
 export class AlertBannerService extends HoistService {
     override xhImpl = true;
 
     static instance: AlertBannerService;
 
-    @managed
-    private timer: Timer;
-
-    get interval(): number {
-        const conf = XH.getConf('xhAlertBannerConfig', {});
-        return conf.enabled && conf.interval ? conf.interval * SECONDS : -1;
-    }
-
-    get enabled(): boolean {
-        return this.interval > 0;
-    }
-
     get lastDismissed(): number {
         return XH.localStorageService.get('xhAlertBanner.lastDismissed');
     }
 
-    override async initAsync() {
-        this.timer = Timer.create({
-            runFn: () => this.checkForBannerAsync(),
-            interval: this.interval
-        });
-    }
-
-    async checkForBannerAsync() {
-        if (!this.enabled) return;
-
-        const data: AlertBannerSpec = await XH.fetchJson({url: 'xh/alertBanner'}),
-            {active, expires, publishDate, message, intent, iconName, enableClose, clientApps} =
-                data,
+    async updateBanner(spec: AlertBannerSpec) {
+        const {active, expires, publishDate, message, intent, iconName, enableClose, clientApps} =
+                spec,
             {lastDismissed, onClose} = this;
 
         if (