+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 68.0.0-SNAPSHOT - unreleased
 
+### 💥 Breaking Changes (upgrade difficulty: 🟢 LOW - Hoist Core update only)
+* Requires `hoist-core >= 21.1`
+
+### ⚙️ Technical
+* Alert Banner is now updated via the shared environment polling for improved simplicity and frequency.
+
 ## 67.0.0 - 2024-09-03
 
 ### 💥 Breaking Changes (upgrade difficulty: 🟢 LOW - Hoist Core update only)